# Implementing Regex optimization on the `MatchNotRegexp` and `MatchNotEqual` matcher type